requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.15
django-filter==21.1
django-import-export==2.7.1
django-nested-admin==3.4.0
django-phonenumber-field==6.0.0
django-phonenumbers==1.0.1
django-quill-editor==0.1.40
django-rest-swagger==2.2.0
django-simple-history==3.0.0
djangorestframework==3.13.0
drf-extensions==0.7.1
drf-writable-nested==0.6.3
drf-yasg==1.20.0
et-xmlfile==1.1.0
FormEncode==2.0.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.8.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.1
Markdown==3.3.6
MarkupPy==1.14
MarkupSafe==2.1.1
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.9
packaging==21.3
phonenumbers==8.12.45
Pillow==9.0.1
psycopg2-binary==2.9.2
pyaes==1.6.1
pyasn1==0.4.8
Pygments==2.11.2
pyparsing==3.0.8
python-monkey-business==1.0.0
python-slugify==6.1.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
slugify==0.0.1
sqlparse==0.4.2
tablib==3.2.0
telebot==0.0.4
telegram==0.0.1
Telethon==1.24.0
text-unidecode==1.3
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
xlrd==2.0.1
xlwt==1.3.0
zipp==3.6.0
